+++ jsonargparse_tests/test_subclasses.py
@@ -548,6 +548,9 @@ def test_subclass_nested_help(parser):
     help_str = get_parse_args_stdout(parser, [f"--op.help={__name__}.Nested", "--op.cal.help=TextCalendar"])
     assert "Help for --op.cal.help=calendar.TextCalendar" in help_str
     assert "--op.cal.firstweekday" in help_str
+    help_str = get_parse_args_stdout(parser, [f"--op.help={__name__}.Nested", "--op.cal.help"])
+    assert "Help for --op.cal.help=calendar.Calendar" in help_str
+    assert "--op.cal.firstweekday" in help_str
 
     with pytest.raises(ArgumentError) as ctx:
         parser.parse_args([f"--op.help={__name__}.Nested", "--op.p1=1"])
@@ -1346,7 +1349,7 @@ def test_add_subclass_required_group(parser):
     parser.add_subclass_arguments(Calendar, "cal", required=True)
     pytest.raises(ArgumentError, lambda: parser.parse_args([]))
     help_str = get_parser_help(parser)
-    assert "[-h] [--cal.help CLASS_PATH_OR_NAME] --cal " in help_str
+    assert "[-h] [--cal.help [CLASS_PATH_OR_NAME]] --cal " in help_str
 
 
 def test_add_subclass_not_required_group(parser):
@@ -1485,6 +1488,7 @@ def test_is_instance_or_supports_protocol(expected, value):
 
 def test_parse_implements_protocol(parser):
     parser.add_argument("--cls", type=Interface)
+    assert "known subclasses:" not in get_parser_help(parser)
     cfg = parser.parse_args([f"--cls={__name__}.ImplementsInterface", "--cls.batch_size=5"])
     assert cfg.cls.class_path == f"{__name__}.ImplementsInterface"
     assert cfg.cls.init_args == Namespace(batch_size=5)
@@ -1492,12 +1496,12 @@ def test_parse_implements_protocol(parser):
     assert isinstance(init.cls, ImplementsInterface)
     assert init.cls.batch_size == 5
     assert init.cls.predict([1.0, 2.0]) == [1.0, 2.0]
-    with pytest.raises(ArgumentError) as ctx:
+    with pytest.raises(ArgumentError, match="is a protocol"):
+        parser.parse_args([f"--cls={__name__}.Interface"])
+    with pytest.raises(ArgumentError, match="does not implement protocol"):
         parser.parse_args([f"--cls={__name__}.NotImplementsInterface1"])
-    ctx.match("does not correspond to a subclass of")
-    with pytest.raises(ArgumentError) as ctx:
+    with pytest.raises(ArgumentError, match="Does not implement protocol Interface"):
         parser.parse_args(['--cls={"batch_size": 5}'])
-    ctx.match("Not a valid subclass of Interface")
 
 
 # callable protocol tests
@@ -1507,7 +1511,7 @@ class CallableInterface(Protocol):
     def __call__(self, items: List[float]) -> List[float]: ...
 
 
-class ImplementsCallableInterface1:
+class ImplementsCallableInterface:
     def __init__(self, batch_size: int):
         self.batch_size = batch_size
 
@@ -1533,8 +1537,8 @@ class NotImplementsCallableInterface3:
 @pytest.mark.parametrize(
     "expected, value",
     [
-        (True, ImplementsCallableInterface1),
-        (False, ImplementsCallableInterface1(1)),
+        (True, ImplementsCallableInterface),
+        (False, ImplementsCallableInterface(1)),
         (False, NotImplementsCallableInterface1),
         (False, NotImplementsCallableInterface2),
         (False, NotImplementsCallableInterface3),
@@ -1543,6 +1547,23 @@ class NotImplementsCallableInterface3:
 )
 def test_implements_callable_protocol(expected, value):
     assert implements_protocol(value, CallableInterface) is expected
+
+
+def test_parse_implements_callable_protocol(parser):
+    parser.add_argument("--cls", type=CallableInterface)
+    assert "known subclasses:" not in get_parser_help(parser)
+    cfg = parser.parse_args([f"--cls={__name__}.ImplementsCallableInterface", "--cls.batch_size=7"])
+    assert cfg.cls.class_path == f"{__name__}.ImplementsCallableInterface"
+    assert cfg.cls.init_args == Namespace(batch_size=7)
+    init = parser.instantiate_classes(cfg)
+    assert isinstance(init.cls, ImplementsCallableInterface)
+    assert init.cls([1.0, 2.0]) == [1.0, 2.0]
+    with pytest.raises(ArgumentError, match="is a protocol"):
+        parser.parse_args([f"--cls={__name__}.CallableInterface"])
+    with pytest.raises(ArgumentError, match="does not implement protocol"):
+        parser.parse_args([f"--cls={__name__}.NotImplementsCallableInterface1"])
+    with pytest.raises(ArgumentError, match="Does not implement protocol CallableInterface"):
+        parser.parse_args(['--cls={"batch_size": 7}'])
 
 
 # parameter skip tests